+++ src/main/java/io/github/devatherock/artifactory/controllers/DockerController.java
@@ -1,12 +1,19 @@
 package io.github.devatherock.artifactory.controllers;
 
 import io.github.devatherock.artifactory.service.DockerBadgeService;
+import io.github.devatherock.artifactory.util.ApiSpecConstants;
 
 import io.micronaut.http.annotation.Controller;
 import io.micronaut.http.annotation.Get;
 import io.micronaut.http.annotation.QueryValue;
 import io.micronaut.scheduling.TaskExecutors;
 import io.micronaut.scheduling.annotation.ExecuteOn;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -25,37 +32,41 @@ public class DockerController {
     private final DockerBadgeService badgeService;
 
     /**
-     * Generates the {@code docker pulls} badge
+     * Generates the docker pulls badge
      * 
      * @param packageName
      * @param badgeLabel
      * @return the pulls badge
      */
     @Get(value = "/pulls", produces = CONTENT_TYPE_BADGE)
-    public String getImagePullCount(@QueryValue("package") String packageName,
-                                    @QueryValue(value = "label", defaultValue = "docker pulls") String badgeLabel) {
+    @Operation(description = "Generates the docker pulls badge", responses = @ApiResponse(description = "An XML representing the SVG docker pulls badge", content = @Content(mediaType = CONTENT_TYPE_BADGE, schema = @Schema(implementation = String.class))))
+    public String getImagePullCount(
+                                    @Parameter(in = ParameterIn.QUERY, description = ApiSpecConstants.API_DESC_PARAM_PACKAGE, example = ApiSpecConstants.EXAMPLE_PARAM_PACKAGE) @QueryValue("package") String packageName,
+                                    @Parameter(in = ParameterIn.QUERY, description = ApiSpecConstants.API_DESC_PARAM_LABEL) @QueryValue(value = "label", defaultValue = "docker pulls") String badgeLabel) {
         LOGGER.debug("In getImagePullCount");
         return badgeService.getPullsCountBadge(packageName, badgeLabel);
     }
 
     /**
-     * Generates the {@code image size} badge
+     * Generates the image size badge
      * 
      * @param packageName
      * @param tag
      * @param badgeLabel
      * @return the image size badge
      */
-    @Get(value = "/image-size", produces = CONTENT_TYPE_BADGE)
-    public String getImageSize(@QueryValue("package") String packageName,
-                               @QueryValue(defaultValue = "latest") String tag,
-                               @QueryValue(value = "label", defaultValue = "image size") String badgeLabel) {
+    @Get(uris = { "/image-size", "image_size" }, produces = CONTENT_TYPE_BADGE)
+    @Operation(description = "Generates the image size badge", responses = @ApiResponse(description = "An XML representing the SVG image size badge", content = @Content(mediaType = CONTENT_TYPE_BADGE, schema = @Schema(implementation = String.class))))
+    public String getImageSize(
+                               @Parameter(in = ParameterIn.QUERY, description = ApiSpecConstants.API_DESC_PARAM_PACKAGE, example = ApiSpecConstants.EXAMPLE_PARAM_PACKAGE) @QueryValue("package") String packageName,
+                               @Parameter(in = ParameterIn.QUERY, description = ApiSpecConstants.API_DESC_PARAM_TAG) @QueryValue(defaultValue = "latest") String tag,
+                               @Parameter(in = ParameterIn.QUERY, description = ApiSpecConstants.API_DESC_PARAM_LABEL) @QueryValue(value = "label", defaultValue = "image size") String badgeLabel) {
         LOGGER.debug("In getImageSize");
         return badgeService.getImageSizeBadge(packageName, tag, badgeLabel);
     }
 
     /**
-     * Generates the {@code layers} badge
+     * Generates the layers badge
      * 
      * @param packageName
      * @param tag
@@ -63,9 +74,11 @@ public class DockerController {
      * @return the layers badge
      */
     @Get(value = "/layers", produces = CONTENT_TYPE_BADGE)
-    public String getImageLayers(@QueryValue("package") String packageName,
-                                 @QueryValue(defaultValue = "latest") String tag,
-                                 @QueryValue(value = "label", defaultValue = "layers") String badgeLabel) {
+    @Operation(description = "Generates the layers badge", responses = @ApiResponse(description = "An XML representing the SVG layers badge", content = @Content(mediaType = CONTENT_TYPE_BADGE, schema = @Schema(implementation = String.class))))
+    public String getImageLayers(
+                                 @Parameter(in = ParameterIn.QUERY, description = ApiSpecConstants.API_DESC_PARAM_PACKAGE, example = ApiSpecConstants.EXAMPLE_PARAM_PACKAGE) @QueryValue("package") String packageName,
+                                 @Parameter(in = ParameterIn.QUERY, description = ApiSpecConstants.API_DESC_PARAM_TAG) @QueryValue(defaultValue = "latest") String tag,
+                                 @Parameter(in = ParameterIn.QUERY, description = ApiSpecConstants.API_DESC_PARAM_LABEL) @QueryValue(value = "label", defaultValue = "layers") String badgeLabel) {
         LOGGER.debug("In getImageLayers");
         return badgeService.getImageLayersBadge(packageName, tag, badgeLabel);
     }