+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - fix(deps): update dependency org.objenesis:objenesis to v3.3
 - fix(deps): update dependency io.micronaut:micronaut-bom to v3.6.1
 - chore(deps): update plugin org.owasp.dependencycheck to v7.1.2
+- Improved the API spec using additional annotations
 
 ## [1.2.0] - 2022-05-15
 ### Added